Connected most public elements to backend for anime

+++ kaguya/models.py
@@ -143,7 +143,7 @@ class Anime(db.Model):
         default=datetime.utcnow)
     title = db.Column(db.String(60), unique=False, nullable=False)
     type = db.Column(db.String(10), nullable=True, default='TV')
-    episodes = db.Column(db.Float, nullable=True, default=0.)
+    episodes = db.Column(db.Integer, nullable=True, default=0)
     rating = db.Column(db.String(50), nullable=True)
     score = db.Column(db.Float, nullable=True, default=0.)
     status = db.Column(db.String(20), default='Finished')
